Refactor sql-db compatibility module, name prefix sync
