+++ sql-db/hibernate/pom.xml
@@ -8,7 +8,7 @@
         <relativePath>../..</relativePath>
     </parent>
     <artifactId>sqldb-hibernate</artifactId>
-    <name>Quarkus QE TS: Hibernate exploratory</name>
+    <name>Quarkus QE TS: SQL Database: Hibernate exploratory</name>
     <packaging>jar</packaging>
     <dependencies>
         <dependency>
