--- sql-db/sql-app-compatibility/src/test/java/io/quarkus/ts/sqldb/sqlapp/DevModeMariadbIT.java
@@ -1,21 +1,0 @@
-package io.quarkus.ts.sqldb.sqlapp;
-
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
-
-import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.QuarkusScenario;
-import io.quarkus.test.services.DevModeQuarkusApplication;
-
-@Tag("QUARKUS-959")
-@QuarkusScenario
-public class DevModeMariadbIT extends AbstractSqlDatabaseIT {
-
-    @DevModeQuarkusApplication
-    static RestService app = new RestService().withProperties("mariadb_app.properties");
-
-    @Test
-    public void mariadbContainerShouldBeStarted() {
-        app.logs().assertContains("Creating container for image: docker.io/mariadb");
-    }
-}
