--- sql-db/sql-app-compatibility/src/test/java/io/quarkus/ts/sqldb/sqlapp/OpenShiftMssqlDatabaseIT.java
@@ -1,11 +1,0 @@
-package io.quarkus.ts.sqldb.sqlapp;
-
-import org.junit.jupiter.api.Disabled;
-
-import io.quarkus.test.scenarios.OpenShiftScenario;
-
-@OpenShiftScenario
-@Disabled("https://github.com/microsoft/mssql-docker/issues/769")
-public class OpenShiftMssqlDatabaseIT extends MssqlDatabaseIT {
-
-}
