--- sql-db/sql-app-compatibility/src/test/java/io/quarkus/ts/sqldb/sqlapp/OpenShiftOracleDatabaseIT.java
@@ -1,10 +1,0 @@
-package io.quarkus.ts.sqldb.sqlapp;
-
-import org.junit.jupiter.api.Disabled;
-
-import io.quarkus.test.scenarios.OpenShiftScenario;
-
-@Disabled("https://github.com/quarkus-qe/quarkus-test-suite/issues/246")
-@OpenShiftScenario
-public class OpenShiftOracleDatabaseIT extends OracleDatabaseIT {
-}
